refactor: 💡 split pagination and date filter and fix styling
